+++ model/infinispan/src/main/java/org/keycloak/models/sessions/infinispan/InfinispanSingleUseTokenStoreProvider.java
@@ -47,7 +47,7 @@ public class InfinispanSingleUseTokenStoreProvider implements SingleUseTokenStor
     }
 
     @Override
-    public boolean putIfAbsent(String tokenId, int lifespanInSeconds) {
+    public boolean putIfAbsent(String tokenId, long lifespanInSeconds) {
         ActionTokenValueEntity tokenValue = new ActionTokenValueEntity(null);
 
         // Rather keep the items in the cache for a bit longer
